调整会员中心部分接口的入参类型

+++ src/main/java/com/zhss/eshop/membership/service/impl/MembershipServiceImpl.java
@@ -44,7 +44,7 @@ public class MembershipServiceImpl implements MembershipService {
 	 * @param totalOrderAmount 订单总金额
 	 * @return 处理结果
 	 */
-	public Boolean informPayOrderEvent(Long userAccountId, Long totalOrderAmount) {
+	public Boolean informPayOrderEvent(Long userAccountId, Double totalOrderAmount) {
 		return true;
 	}
 	
@@ -54,7 +54,7 @@ public class MembershipServiceImpl implements MembershipService {
 	 * @param totalOrderAmount 订单总金额
 	 * @return 处理结果
 	 */
-	public Boolean informFinishReturnGoodsEvent(Long userAccountId, Long totalOrderAmount) {
+	public Boolean informFinishReturnGoodsEvent(Long userAccountId, Double totalOrderAmount) {
 		return true;
 	}
